Homework2 done

+++ src/main/java/ru/hogwarts/school/service/StudentServiceImpl.java
@@ -1,6 +1,9 @@
 package ru.hogwarts.school.service;
 
 import java.util.Collection;
+import java.util.Optional;
+
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import ru.hogwarts.school.model.Student;
 import ru.hogwarts.school.repository.StudentRepository;
@@ -25,18 +28,50 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Student findStudent(long id) {
-        return studentRepository.findById(id).orElse(null);
+    public ResponseEntity<Student> findStudent(long id) {
+// Проверка некорректного ID
+        if (id <= 0) {
+            return ResponseEntity.badRequest().build();
+        }
+
+        Optional<Student> studentOptional = studentRepository.findById(id);
+
+        if (studentOptional.isPresent()) {
+            return ResponseEntity.ok(studentOptional.get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @Override
-    public Student deleteStudent(long id) {
+    public ResponseEntity<Void> deleteStudent(long id) {
+// Проверка на недопустимые значения id
+        if (id <= 0) {
+            return ResponseEntity.badRequest().build(); // 400 Bad Request
+        }
+
+        Optional<Student> studentOptional = studentRepository.findById(id);
+        if (!studentOptional.isPresent()) {
+            return ResponseEntity.notFound().build(); // 404 Not Found
+        }
+
         studentRepository.deleteById(id);
-        return null;
+        return ResponseEntity.noContent().build(); // 204 No Content
+    }
+
+    @Override
+    public Collection<Student> getAllStudents() {
+        return studentRepository.findAll();
     }
 
     @Override
     public Collection<Student> findByAge(int age) {
         return studentRepository.findByAge(age);
+    }
+
+
+    @Override
+    public Collection<Student> findByAgeBetween(int minAge, int maxAge) {
+        return studentRepository.findByAgeBetween(minAge, maxAge);
     }
 }
